backend/requirements.txt
asgiref==3.8.1
Django==4.2.18
django-cors-headers==4.6.0
djangorestframework==3.15.2
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
